Refactored GlycanDetail
